Changed generic view slice to reducer
