+++ libs/core/__deprecated__/renderer/store/reducers.ts
@@ -18,7 +18,7 @@ import { templateReducer } from "Core/templates/reducers"
 import { newsReducer } from "Core/news/reducers"
 import { settingsReducer } from "Core/settings/reducers"
 import { updateOsReducer } from "Core/update/reducers"
-import { genericViewsSlice } from "generic-view/store"
+import { genericViewsReducer } from "generic-view/store"
 
 export const reducers = {
   device: deviceReducer,
@@ -35,7 +35,7 @@ export const reducers = {
   news: newsReducer,
   settings: settingsReducer,
   update: updateOsReducer,
-  genericViews: genericViewsSlice.reducer,
+  genericViews: genericViewsReducer,
 }
 
 export const combinedReducers = combineReducers(reducers)
