+++ libs/generic-view/store/src/index.ts
@@ -3,5 +3,6 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./lib/slice"
 export * from "./lib/actions"
+export * from "./lib/reducer"
+export * from "./lib/generic-dispatch-demo"
