--- libs/generic-view/store/src/lib/actions/index.ts
@@ -1,6 +1,0 @@
-/**
- * Copyright (c) Mudita sp. z o.o. All rights reserved.
- * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
- */
-
-export * from "./generic-dispatch-demo"
