+++ libs/generic-view/store/src/lib/generic-dispatch-demo.ts
@@ -14,9 +14,10 @@ import {
 } from "generic-view/views"
 import { useDispatch } from "react-redux"
 import { useEffect } from "react"
-import { menuConfig } from "../../../../../../demo-data/demo-menu"
+import { menuConfig } from "../../../../../demo-data/demo-menu"
 import { generateMenu } from "generic-view/utils"
-import { genericViewsSlice } from "../slice"
+import { setMenu, setViewData, setViewLayout } from "./actions"
+import { genericViewsReducer } from "./reducer"
 
 // For demo purposes to simulate device connection and async data fetching
 export const useGenericStoreDemo = () => {
@@ -27,18 +28,18 @@ export const useGenericStoreDemo = () => {
       return
     }
     void (async () => {
-      if (genericViewsSlice.getInitialState().menu) {
+      if (genericViewsReducer.getInitialState().menu) {
         return
       }
 
       // Simulate menu setup after device connection
       await wait(2000)
-      dispatch(genericViewsSlice.actions.setMenu(generateMenu(menuConfig)))
+      dispatch(setMenu(generateMenu(menuConfig)))
 
       // Simulate mc-overview layout
       await wait(1000)
       dispatch(
-        genericViewsSlice.actions.setViewLayout({
+        setViewLayout({
           feature: "mc-overview",
           layout: generateMcOverviewLayout(mcOverviewConfig),
         })
@@ -47,7 +48,7 @@ export const useGenericStoreDemo = () => {
       // Simulate mc-calendar layout
       await wait(1000)
       dispatch(
-        genericViewsSlice.actions.setViewLayout({
+        setViewLayout({
           feature: "mc-calendar",
           layout: generateMcCalendarLayout(mcCalendarConfig),
         })
@@ -55,7 +56,7 @@ export const useGenericStoreDemo = () => {
 
       await wait(1000)
       dispatch(
-        genericViewsSlice.actions.setViewData({
+        setViewData({
           feature: "mc-overview",
           data: mcOverviewDemoData,
         })
